+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatMenuUI.java
@@ -111,6 +111,7 @@ public class FlatMenuUI
 	protected void uninstallDefaults() {
 		super.uninstallDefaults();
 
+		FlatMenuItemRenderer.clearClientProperties( menuItem.getParent() );
 		renderer = null;
 		oldStyleValues = null;
 	}