+++ flatlaf-testing/src/main/java/com/formdev/flatlaf/testing/FlatMenusTest.java
@@ -165,6 +165,11 @@ public class FlatMenusTest
 		JMenu menu12 = new JMenu();
 		JMenuItem menuItem41 = new JMenuItem();
 		JMenuItem menuItem42 = new JMenuItem();
+		JMenuItem menuItem43 = new JMenuItem();
+		JMenuItem menuItem44 = new JMenuItem();
+		JMenuItem menuItem45 = new JMenuItem();
+		JMenuItem menuItem46 = new JMenuItem();
+		JMenuItem menuItem47 = new JMenuItem();
 		menuBar2 = new JMenuBar();
 		JMenu menu8 = new JMenu();
 		FlatMenusTest.LargerMenuItem menuItem13 = new FlatMenusTest.LargerMenuItem();
@@ -397,6 +402,32 @@ public class FlatMenusTest
 				menuItem42.setDisabledIcon(new ImageIcon(getClass().getResource("/com/formdev/flatlaf/testing/disabled_icons_test/intellij-menu-paste.png")));
 				menuItem42.setEnabled(false);
 				menu12.add(menuItem42);
+				menu12.addSeparator();
+
+				//---- menuItem43 ----
+				menuItem43.setText("text");
+				menuItem43.setIcon(new ImageIcon(getClass().getResource("/com/formdev/flatlaf/testing/test16.png")));
+				menu12.add(menuItem43);
+
+				//---- menuItem44 ----
+				menuItem44.setText("text");
+				menuItem44.setIcon(new ImageIcon(getClass().getResource("/com/formdev/flatlaf/testing/test24.png")));
+				menu12.add(menuItem44);
+
+				//---- menuItem45 ----
+				menuItem45.setText("text");
+				menuItem45.setIcon(new ImageIcon(getClass().getResource("/com/formdev/flatlaf/testing/test32.png")));
+				menu12.add(menuItem45);
+
+				//---- menuItem46 ----
+				menuItem46.setText("text");
+				menuItem46.setIcon(new ImageIcon(getClass().getResource("/com/formdev/flatlaf/testing/test48.png")));
+				menu12.add(menuItem46);
+
+				//---- menuItem47 ----
+				menuItem47.setText("text");
+				menuItem47.setIcon(new ImageIcon(getClass().getResource("/com/formdev/flatlaf/testing/test64.png")));
+				menu12.add(menuItem47);
 			}
 			menuBar1.add(menu12);
 		}